fix controllers structure & naming
