+++ registration/src/main/java/com/team1/registration/configs/WebSecurityConfig.java
@@ -2,6 +2,7 @@ package com.team1.registration.configs;
 
 
 import com.team1.registration.services.UserService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
@@ -17,6 +18,7 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
 
     private final UserService userService;
 
+    @Autowired
     public WebSecurityConfig(UserService userService) {
         this.userService = userService;
     }
