+++ registration/src/main/java/com/team1/registration/controllers/navbar/LoginPageController.java
@@ -5,8 +5,5 @@ import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
 public class LoginPageController {
-    @GetMapping("/login")
-    public String login(){
-        return "navbar/login";
-    }
+
 }
