+++ registration/src/main/java/com/team1/registration/controllers/navbar/TournamentPageController.java
@@ -5,8 +5,5 @@ import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
 public class TournamentPageController {
-    @GetMapping("tournaments")
-    public String tournaments(){ // todo: fix name
-        return "tournaments";
-    }
+
 }
